+++ ebpf/python/pyperf.go
@@ -60,27 +60,18 @@ func NewPerf(logger log.Logger, metrics *metrics.PythonMetrics, perfEventMap *eb
 
 // StartPythonProfiling return python init error, second return value is true if process is still alive,
 // third return value - python process data
-func (s *Perf) StartPythonProfiling(pid uint32, serviceName string) (*PerfPyPidData, bool, error) {
+func (s *Perf) StartPythonProfiling(pid uint32, data *PerfPyPidData, serviceName string) error {
 	if s.pidCache.Contains(pid) {
-		return nil, false, nil
+		return nil
 	}
-	data, err := GetPyPerfPidData(s.logger, pid)
-	if err != nil {
-		alive := processAlive(pid)
-		if alive {
-			s.metrics.PidDataError.WithLabelValues(serviceName).Inc()
-		}
-		s.pidCache.Add(pid, nil) // to never try again
-		return nil, alive, fmt.Errorf("error collecting python data %w", err)
-	}
-	err = s.pidDataHashMap.Update(pid, data, ebpf.UpdateAny)
+
+	err := s.pidDataHashMap.Update(pid, data, ebpf.UpdateAny)
 	if err != nil { // should never happen
-		s.pidCache.Add(pid, nil) // to never try again
-		return nil, true, fmt.Errorf("updating pid data hash map: %w", err)
+		return fmt.Errorf("updating pid data hash map: %w", err)
 	}
 	s.metrics.ProcessInitSuccess.WithLabelValues(serviceName).Inc()
 	s.pidCache.Add(pid, data)
-	return data, true, nil
+	return nil
 }
 
 func (s *Perf) loop() {
@@ -270,9 +261,4 @@ func (s *LazySymbols) getSymbol(id uint32, svc string) (*PerfPySymbol, error) {
 		return symbol, nil
 	}
 	return nil, fmt.Errorf("symbol %d not found", id)
-}
-
-func processAlive(pid uint32) bool {
-	_, err := os.Stat(fmt.Sprintf("/proc/%d", pid))
-	return err == nil
 }
